+++ src/OpenDDD/Application/Settings/Auth/RbacProvider.cs
@@ -4,6 +4,7 @@
 	{
 		None,
 		Negative,
+		Positive,
 		PowerIAM
 	}
 }