+++ src/OpenDDD/Application/Settings/SettingsException.cs
@@ -1,16 +1,15 @@
 ﻿using System;
 using OpenDDD.Application.Error;
-using ApplicationException = OpenDDD.Application.Error.ApplicationException;
 
 namespace OpenDDD.Application.Settings
 {
     public class SettingsException : Error.ApplicationException
     {
         public static SettingsException Invalid(string spec)
-            => new SettingsException(ApplicationError.Settings_InvalidSettings(spec));
+            => new SettingsException(ApplicationError.Settings_Invalid(spec));
         
         public static SettingsException Invalid(string spec, Exception inner)
-            => new SettingsException(ApplicationError.Settings_InvalidSettings(spec), inner);
+            => new SettingsException(ApplicationError.Settings_Invalid(spec), inner);
         
         public static SettingsException AuthEnabledButNoPrivateKey()
             => new SettingsException(ApplicationError.Settings_AuthEnabledButNoPrivateKey());