+++ src/OpenDDD/Infrastructure/Ports/Adapters/PubSub/ServiceBus/ServiceBusDomainEventAdapter.cs
@@ -12,7 +12,7 @@ namespace OpenDDD.Infrastructure.Ports.Adapters.PubSub.ServiceBus
 			ISettings settings,
 			ILogger logger,
 			IMonitoringPort monitoringAdapter,
-			SerializerSettings serializerSettings) :
+			ConversionSettings conversionSettings) :
 			base(
 				settings.General.Context,
 				settings.General.Context,
@@ -21,7 +21,7 @@ namespace OpenDDD.Infrastructure.Ports.Adapters.PubSub.ServiceBus
 				settings.Azure.ServiceBus.SubName,
 				logger,
 				monitoringAdapter,
-				serializerSettings)
+				conversionSettings)
 		{
 			
 		}